# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School ID App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scan:
    post:
      operationId: processScan
      tags: [scan]
      summary: Process a scan event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanBody"
      responses:
        "200":
          description: Scan processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "404":
          description: Student not found
  /scan/events:
    get:
      operationId: listScanEvents
      tags: [scan]
      summary: List scan events
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scan events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanEvent"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get live dashboard summary
      parameters:
        - name: grade
          in: query
          schema:
            type: string
          description: Filter by grade (e.g. "8", "9", "10", "11", "12")
        - name: className
          in: query
          schema:
            type: string
          description: Filter by class name (e.g. "8A", "10B")
      responses:
        "200":
          description: Dashboard summary with KPIs, exceptions, and feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/feed:
    get:
      operationId: getDashboardFeed
      tags: [dashboard]
      summary: Get live activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Live activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedItem"
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        grade:
          type: string
        className:
          type: string
        photoUrl:
          type: ["string", "null"]
        qrCode:
          type: string
        isActive:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - firstName
        - lastName
        - grade
        - className
        - qrCode
        - isActive
        - createdAt
        - updatedAt
    StudentWithState:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            currentState:
              type: string
            lastSeenAt:
              type: ["string", "null"]
            lastSeenLocation:
              type: ["string", "null"]
          required:
            - currentState
...
    ScanType:
      type: string
      enum:
        - gate_in
        - gate_out
        - checkout
        - class
        - event
        - assembly
        - activity
        - detention
        - club
    ScanBody:
      type: object
      properties:
        qrCode:
          type: string
        scanType:
          $ref: "#/components/schemas/ScanType"
        location:
          type: ["string", "null"]
        activityId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - qrCode
        - scanType
    ScanResult:
      type: object
      properties:
        scanEvent:
          $ref: "#/components/schemas/ScanEvent"
        student:
          $ref: "#/components/schemas/StudentWithState"
        warnings:
          type: array
          items:
            type: string
      required:
        - scanEvent
        - student
        - warnings
    ScanEvent:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        scannedById:
          type: ["integer", "null"]
        scanType:
          $ref: "#/components/schemas/ScanType"
        location:
          type: ["string", "null"]
        activityId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        studentName:
          type: ["string", "null"]
      required:
        - id
        - studentId
        - scanType
        - createdAt
    DashboardSummary:
      type: object
      properties:
        kpis:
          $ref: "#/components/schemas/DashboardKpis"
        exceptions:
          $ref: "#/components/schemas/DashboardExceptions"
        statusDistribution:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        recentFeed:
          type: array
          items:
            $ref: "#/components/schemas/FeedItem"
        lastUpdated:
          type: string
        availableClassesByGrade:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: "Mapping of grade to available class names for filter chip population"
      required:
        - kpis
        - exceptions
        - statusDistribution
        - recentFeed
        - lastUpdated
        - availableClassesByGrade
    DashboardKpis:
      type: object
      properties:
        total:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        checkedOut:
          type: integer
        unaccounted:
          type: integer
        onCampus:
          type: integer
        inClass:
          type: integer
        atEvent:
          type: integer
      required:
        - total
        - present
        - absent
        - late
        - checkedOut
        - unaccounted
        - onCampus
        - inClass
        - atEvent
    DashboardExceptions:
      type: object
      properties:
        missingFromClass:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithState"
        unaccountedStudents:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithState"
        lateArrivals:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithState"
        checkedOutWithoutReason:
          type: array
          items:
            $ref: "#/components/schemas/StudentWithState"
      required:
        - missingFromClass
        - unaccountedStudents
        - lateArrivals
        - checkedOutWithoutReason
    StatusCount:
      type: object
      properties:
        state:
          type: string
        count:
          type: integer
      required:
        - state
        - count
    FeedItem:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        studentName:
          type: string
        scanType:
          $ref: "#/components/schemas/ScanType"
        createdAt:
          type: string
        studentId:
          type: integer
      required:
        - id
        - message
        - studentName
        - scanType
        - createdAt
        - studentId
...